backend/requirements.txt
fastapi[standard]==0.129.0
uvicorn[standard]>=0.30.0
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.30.0
alembic>=1.13.0
pydantic>=2.9.0
pydantic-settings>=2.5.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
pytest>=8.3.0
pytest-asyncio>=0.24.0
httpx>=0.27.0
aiosqlite>=0.20.0
